Added function so the future dates are accurate
This helps for when the numbers of days in the future date strings pass the number of days in an actual month. It will return to the current date and just add the number of days needed to a new month. Also helps for the dates to be in conjunction with the weather API time data.